# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 小眠 sleep-api
  version: 0.1.0
  description: P0~P3 — Auth + AI BFF + Quota + Sync + Energy + Social + Push

servers:
  - url: http://localhost:8787

paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK

  /auth/guest:
    post:
      summary: Create guest session
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceId:
                  type: string
      responses:
        '200':
          description: Guest token

  /auth/login:
    post:
      summary: Login with username/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User token
        '401':
          description: Invalid credentials

  /ai/quota:
    get:
      summary: Daily AI quota snapshot
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Quota

  /ai/chat:
    post:
      summary: Chat with XiaoMian
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                history:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        enum: [user, assistant]
                      content:
                        type: string
                systemPrompt:
                  type: string
                fallback:
                  type: string
                temperature:
                  type: number
                maxTokens:
                  type: integer
      responses:
        '200':
          description: Chat result
        '429':
          description: Quota exceeded

  /ai/dream/interpret:
    post:
      summary: Structured dream interpretation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - dreamText
                - mood
                - contextBlock
                - personalImagery
                - isIncomplete
              properties:
                dreamText:
                  type: string
                mood:
                  type: string
                userGuess:
                  type: string
                contextBlock:
                  type: string
                personalImagery:
                  type: array
                  items:
                    type: object
                    properties:
                      motif:
                        type: string
                      personalNote:
                        type: string
                bottleSnippet:
                  type: string
                isIncomplete:
                  type: boolean
                systemPrompt:
                  type: string
      responses:
        '200':
          description: Interpret result
        '429':
          description: Quota exceeded


  # ========== P3: Social ==========

  /social/friends:
    get:
      summary: List friends
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Friends list
          content:
            application/json:
              schema:
                type: object
                properties:
                  friends:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        username: { type: string }
                        nickname: { type: string }
        '403':
          description: Guest not allowed

  /social/friends/pending:
    get:
      summary: List pending friend requests
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pending requests

  /social/friends/request:
    post:
      summary: Send friend request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [friendId]
              properties:
                friendId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Request sent
        '409':
          description: Already requested or exists

  /social/friends/accept:
    post:
      summary: Accept friend request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [friendId]
              properties:
                friendId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Accepted
        '404':
          description: No pending request

  /social/friends/remove:
    post:
      summary: Remove friend
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [friendId]
              properties:
                friendId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Removed

  /social/bottles:
    post:
      summary: Throw a dream bottle
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content, bottleType]
              properties:
                content:
                  type: string
                  maxLength: 2000
                moodTags:
                  type: array
                  items:
                    type: string
                    maxLength: 20
                bottleType:
                  type: string
                  enum: [random, directed]
                recipientId:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Bottle created
        '400':
          description: Recipient required for directed bottle
        '403':
          description: Guest not allowed

  /social/bottles/random:
    get:
      summary: Pick a random bottle
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Bottle picked
        '404':
          description: No bottles available

  /social/bottles/sent:
    get:
      summary: List bottles I sent
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sent bottles

  /social/bottles/received:
    get:
      summary: List bottles I received
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Received bottles

  /social/bottles/{id}/reply:
    post:
      summary: Reply to a bottle
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  maxLength: 2000
      responses:
        '200':
          description: Reply created
        '404':
          description: Bottle not found

  /social/bottles/{id}/read:
    post:
      summary: Mark a bottle as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Marked as read
        '404':
          description: Bottle not found

  /social/bottles/{id}/replies:
    get:
      summary: List replies to a bottle
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Replies list

  /social/feed:
    get:
      summary: Browse feed posts
      security:
        - bearerAuth: []
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Feed posts with nextCursor
    post:
      summary: Create a feed post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, content]
              properties:
                type:
                  type: string
                  enum: [milestone, dream_share, check_in, achievement]
                content:
                  type: object
      responses:
        '201':
          description: Post created

  /social/feed/{id}/like:
    post:
      summary: Like/unlike a feed post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Like toggled

  # ========== P3: Push ==========

  /push/register:
    post:
      summary: Register push device token
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform, token]
              properties:
                platform:
                  type: string
                  enum: [android, ios]
                token:
                  type: string
      responses:
        '200':
          description: Device registered

  /push/unregister:
    post:
      summary: Unregister push device token
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform, token]
              properties:
                platform:
                  type: string
                  enum: [android, ios]
                token:
                  type: string
      responses:
        '200':
          description: Device unregistered

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT